requirements.txt
gradio>=4.0.0
chatterbox-tts
torch
torchaudio
numpy
librosa
soundfile
